Added new material + Changed Render Texture Settings + Updated VideoPlayback.cs
Added new material
Added loadingMat_Unlit Material which is applied while the Video is Loading

Changed Render Texture Settings
Changed the Render Texture to use R32G32B32A32_UINT Colour Format to reduce Colour Banding

Updated VideoPlayback.cs
Updated the Script to have additional Checks and Functionality. It can now take in an AudioSource and use that for AudioPlayback instead of Direct Playback. But if it is not detected, it will switch to Direct Playback by Default.